+++ cell/CellTri3.h
@@ -14,10 +14,10 @@
 template<int TRI_N>
 class CellTri3 : public CellTri<TRI_N>, public Tri3 {
 public:
-	static const UInt np = (TRI_N+1) * (TRI_N+2) / 2;
-	static const UInt nfp = (TRI_N + 1);
-	static const UInt faces = 3;
-	static const UInt vertices = 3;
+	static const size_t np = (TRI_N+1) * (TRI_N+2) / 2;
+	static const size_t nfp = (TRI_N + 1);
+	static const size_t faces = 3;
+	static const size_t vertices = 3;
 	CellTri3(
 	 const Tri3& base);
 	virtual
